use double precision floats in storage esp

+++ src/main/java/meteordevelopment/meteorclient/utils/render/IVertexConsumerProvider.java
@@ -8,5 +8,5 @@ package meteordevelopment.meteorclient.utils.render;
 import net.minecraft.client.render.VertexConsumerProvider;
 
 public interface IVertexConsumerProvider extends VertexConsumerProvider {
-    void setOffset(float offsetX, float offsetY, float offsetZ);
+    void setOffset(double offsetX, double offsetY, double offsetZ);
 }
